cloud/requirements.txt
# AWS Lambda Architecture Dependencies

# Core AWS SDK
boto3==1.35.84
botocore==1.35.84

# Data Processing
pandas==2.2.3
polars==1.29.0
numpy==2.2.4
pyarrow==19.0.1

# Time Series & Financial Data
polygon-api-client==1.14.4
yfinance==0.2.55

# Database Clients
psycopg2-binary==2.9.10
redis==5.0.1
duckdb==1.3.0
# Stream Processing
kafka-python==2.1.4

# API Framework
fastapi==0.115.12
uvicorn==0.34.0
pydantic==2.11.1

# HTTP Clients
httpx==0.28.1
aiohttp==3.11.14
websockets==14.2

# Time & Date Handling
python-dateutil==2.9.0.post0
pytz==2025.2
pendulum==3.0.0

# Configuration & Environment
python-dotenv==1.1.0
pydantic-settings==2.8.1

# Monitoring & Logging
structlog==24.1.0

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
moto==5.0.21  # AWS mocking for tests

# Development Tools
black==24.10.0
ruff==0.8.4

# AWS CDK (optional - install separately if needed)
# aws-cdk-lib==2.167.1  # Install via: npm install -g aws-cdk
# constructs==10.4.2     # Install via: npm install -g aws-cdk

# Local Development
localstack==3.8.1

# Financial Analysis
ta==0.11.0

# Async Programming
# asyncio is part of Python standard library (no need to install)
aiofiles==24.1.0

# Note: serverless framework is a Node.js package, install via:
# npm install -g serverless
#
# Note: docker-compose is a standalone tool, install via:
# https://docs.docker.com/compose/install/
